updated the structure
